avoid loading loading as many seml modules as possible when completing
