+++ pyproject.toml
@@ -71,7 +71,11 @@ dev-dependencies = [
     "build>=1.2.1",
     "twine>=5.1.1",
     "backports-tarfile>=1.2.0",
+    "tuna>=0.5.11",
 ]
+
+[tool.uv.workspace]
+members = ["new_stuff"]
 
 
 [tool.ruff.format]
