+++ src/seml/cli_utils/__init__.py
@@ -1,0 +1,4 @@
+from .cache import cache_to_disk
+from .module_hider import AUTOCOMPLETING, ModuleHider
+
+__all__ = ['cache_to_disk', 'AUTOCOMPLETING', 'ModuleHider']
