+++ src/seml/cli_utils/module_hider.py
