+++ src/seml/cli_utils/typer.py
@@ -1,4 +1,4 @@
-from seml.utils.module_hider import ModuleHider
+from seml.cli_utils.module_hider import ModuleHider
 
 # If we are in a completion shell, we don't want to import rich.
 # due to its high import time.
