+++ src/seml/console/__init__.py
@@ -13,7 +13,7 @@ from rich.console import Console
 from rich.padding import Padding
 from rich.rule import Rule
 
-from seml.utils.typer import prompt as typer_prompt
+from seml.cli_utils.typer import prompt as typer_prompt
 
 try:
     terminal_width = os.get_terminal_size().columns
